make migrations for service

+++ database/migrations/2023_07_10_163123_create_transactions_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->string('customer');
-            $table->decimal('berat');
-            $table->decimal('total_harga');
+            $table->double('berat');
+            $table->double('total_harga');
             $table->date('tgl_masuk');
             $table->date('tgl_keluar');
             $table->boolean('is_complete')->default(0);
